+++ configurations/pvconfiguration.h
@@ -14,8 +14,8 @@ class PvConfiguration
 {
     Q_GADGET
     Q_PROPERTY(QUuid pvThingId READ pvThingId WRITE setPvThingId)
-    Q_PROPERTY(double longitude READ longitude WRITE setLongitude USER true)
-    Q_PROPERTY(double latitude READ latitude WRITE setLatitude USER true)
+    Q_PROPERTY(float longitude READ longitude WRITE setLongitude USER true)
+    Q_PROPERTY(float latitude READ latitude WRITE setLatitude USER true)
     Q_PROPERTY(double roofPitch READ roofPitch WRITE setRoofPitch USER true)
     Q_PROPERTY(double alignment READ alignment WRITE setAlignment USER true)
     Q_PROPERTY(double kwPeak READ kwPeak WRITE setKwPeak USER true)
@@ -26,11 +26,11 @@ public:
     ThingId pvThingId() const;
     void setPvThingId(const ThingId &pvThingId);
 
-    double longitude() const;
-    void setLongitude(const double longitude);
+    float longitude() const;
+    void setLongitude(const float longitude);
 
-    double latitude() const;
-    void setLatitude(const double latitude);
+    float latitude() const;
+    void setLatitude(const float latitude);
 
     int roofPitch() const;
     void setRoofPitch(const int roofPitch);
@@ -48,8 +48,8 @@ public:
 
 private:
     ThingId m_pvThingId;
-    double m_latitude = 0;
-    double m_longitude = 0;
+    float m_latitude = 0;
+    float m_longitude = 0;
     int m_roofPitch = 0;
     int m_alignment = 0;
     float m_kwPeak = 0;
